Inline getMediaTagNames into useBrowserInit, remove prefilterByMediaTags.ts
